+++ cf/read_write/netcdf/netcdfwrite.py
@@ -103,8 +103,15 @@ class NetCDFWrite(cfdm.read_write.netcdf.NetCDFWrite):
 
                         raise ValueError(
                             f"Can't write {cfvar!r} as a CFA-netCDF "
-                            "aggregation variable. Consider setting "
-                            "cfa={'strict': False}"
+                            "aggregation variable. If the variable was read "
+                            "from disk then setting chunks=None as an "
+                            "argument to cf.read will likely solve the "
+                            "problem. "
+                            "Alternatively, you could consider setting "
+                            "cfa={'strict': False} as an argument to "
+                            "cf.write, but note the this will create a copy "
+                            "of the data for this variable in the output "
+                            "dataset."
                         )
 
                     return cfa_get_write
@@ -464,7 +471,7 @@ class NetCDFWrite(cfdm.read_write.netcdf.NetCDFWrite):
         ):
             f_ncdim = f"f_{ncdim}"
             if f_ncdim not in g["dimensions"]:
-                # Create a new fragement dimension
+                # Create a new fragment dimension
                 self._write_dimension(f_ncdim, None, size=size)
 
             fragment_ncdimensions.append(f_ncdim)
@@ -566,55 +573,6 @@ class NetCDFWrite(cfdm.read_write.netcdf.NetCDFWrite):
                 "aggregated_dimensions": " ".join(ncdimensions),
                 "aggregated_data": " ".join(sorted(aggregated_data_attr)),
             },
-        )
-
-    def _convert_to_builtin_type(self, x):
-        """Convert a non-JSON-encodable object to a JSON-encodable
-        built-in type.
-
-        Possible conversions are:
-
-        ==============  =============  ======================================
-        Input object    Output object  numpy data types covered
-        ==============  =============  ======================================
-        numpy.bool_     bool           bool
-        numpy.integer   int            int, int8, int16, int32, int64, uint8,
-                                       uint16, uint32, uint64
-        numpy.floating  float          float, float16, float32, float64
-        ==============  =============  ======================================
-
-        .. versionadded:: 3.0.0
-
-        :Parameters:
-
-            x:
-
-        :Returns:
-
-            'int' or `float` or `bool`
-
-        **Examples:**
-
-        >>> type(_convert_to_builtin_type(numpy.bool_(True)))
-        bool
-        >>> type(_convert_to_builtin_type(numpy.array([1.0])[0]))
-        double
-        >>> type(_convert_to_builtin_type(numpy.array([2])[0]))
-        int
-
-        """
-        if isinstance(x, np.bool_):
-            return bool(x)
-
-        if isinstance(x, np.integer):
-            return int(x)
-
-        if isinstance(x, np.floating):
-            return float(x)
-
-        raise TypeError(
-            f"{type(x)!r} object can't be converted to a JSON serializable "
-            f"type: {x!r}"
         )
 
     def _check_valid(self, array, cfvar=None, attributes=None):
@@ -908,16 +866,14 @@ class NetCDFWrite(cfdm.read_write.netcdf.NetCDFWrite):
                 if file_details:
                     raise ValueError(
                         "Can't write CFA-netCDF aggregation variable from "
-                        f"{cfvar!r} when the "
-                        f"dask storage chunk defined by indices {indices} "
-                        "spans two or more files"
+                        f"{cfvar!r}: Dask storage chunk defined by indices "
+                        f"{indices} spans two or more fragment files"
                     )
 
                 raise ValueError(
                     "Can't write CFA-netCDF aggregation variable from "
-                    f"{cfvar!r} when the "
-                    f"dask storage chunk defined by indices {indices} spans "
-                    "zero files"
+                    f"{cfvar!r}: Dask storage chunk defined by indices "
+                    f"{indices} spans zero files"
                 )
 
             filenames, addresses, formats = file_details.pop()